Adding python stub prog to run battle ship using bin rep.
The board is a binary string where the ships pos are set to 1.
moves are binary strings. The hits are calclated by `&`ing the
move with the board. We keep track of hits with another binary
string which is initially set to all 0 and then `|`'d with
a successful move.